--- dea.server/src/main/java/fr/ynov/dap/Web/package-info.java
@@ -1,9 +1,0 @@
-/**
- * package "google"
- */
-/**
- * @author antod
- *
- */
-
-package fr.ynov.dap.Web;
