+++ dea.server/src/main/java/fr/ynov/dap/web/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * package "google"
+ */
+/**
+ * @author antod
+ *
+ */
+
+package fr.ynov.dap.web;
